chore: Update npm dependencies
The package.json and package-lock.json files have been modified to update the dependencies. The axios and express packages have been updated to their latest versions, and the cors package has been added as a new dependency. This update is necessary to ensure compatibility and take advantage of the latest features and bug fixes provided by these packages.